requirements.txt
-i https://pypi.org/simple
amqp==5.3.1; python_version >= '3.6'
anyio==4.8.0; python_version >= '3.9'
asgiref==3.9.1; python_version >= '3.9'
beautifulsoup4==4.13.3; python_full_version >= '3.7.0'
billiard==4.2.1; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
certifi==2025.8.3; python_version >= '3.7'
charset-normalizer==3.4.3; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
crispy-bootstrap5==2024.10; python_version >= '3.8'
diff-match-patch==20241021; python_version >= '3.7'
django==4.2.23; python_version >= '3.8'
django-bootstrap-v5==1.0.11; python_version >= '3.6' and python_version < '4.0'
django-crispy-forms==2.3; python_version >= '3.8'
django-environ==0.12.0; python_version >= '3.9' and python_version < '4'
django-import-export==4.3.9; python_version >= '3.9'
django-paystack==1.0.0; python_version >= '3.9'
django-request==1.7.0; python_version >= '3.8'
django-widget-tweaks==1.5.0; python_version >= '3.8'
et-xmlfile==2.0.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
kombu==5.4.2; python_version >= '3.8'
openpyxl==3.1.5; python_version >= '3.8'
paystack-api==0.1.2; python_version >= '3.6'
pillow==11.1.0; python_version >= '3.9'
prompt-toolkit==3.0.50; python_full_version >= '3.8.0'
psycopg==3.2.5; python_version >= '3.8'
pypaystack2==2.0.3; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
rapidfuzz==3.12.2; python_version >= '3.9'
redis==5.2.1; python_version >= '3.8'
requests==2.32.5; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
tablib==3.8.0; python_version >= '3.9'
typing-extensions==4.12.2; python_version < '3.13'
tzdata==2025.1; python_version >= '2'
urllib3==2.5.0; python_version >= '3.9'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
